requirements.txt
braintree==3.57.1
frappe
gocardless-pro==1.11.0
googlemaps==3.1.1
pandas==0.24.2
plaid-python==6.0.0
PyGithub==1.44.1
python-stdnum==1.12
ShopifyAPI==8.0.4
Unidecode==1.1.1
WooCommerce==2.1.1
